Holy crappers shaggie, it now sends structs!
Fancy conversion things on unity side, now utilizes a struct that contains all important information!

+++ TestCPPLibrary/TestCPPLibrary/TestCPPLibrary.h
@@ -3,7 +3,7 @@
 #include "stdafx.h"
 extern "C" {
 	__declspec(dllexport) bool Connect(int ip0, int ip1, int ip2, int ip3, int port);
-	__declspec(dllexport) bool SendPosition(float* data);
+	__declspec(dllexport) bool SendPosition(int* data);
 	__declspec(dllexport) bool ReceiveInformation();
 	void Receive();
 }
